Move usage-emails to usage package / clean shared package
