+++ packages/account-usage/tsconfig.json
@@ -16,6 +16,12 @@
         },
         {
             "path": "../database"
+        },
+        {
+            "path": "../shared"
+        },
+        {
+            "path": "../email"
         }
     ],
     "include": ["lib/**/*", "../utils/lib/vitest.d.ts"]
