+++ packages/server/lib/controllers/v1/account/getEmailByExpiredToken.ts
@@ -1,8 +1,9 @@
 import * as z from 'zod';
 
-import { sendVerificationEmail, userService } from '@nangohq/shared';
+import { userService } from '@nangohq/shared';
 import { getLogger, requireEmptyQuery, zodErrorToHTTP } from '@nangohq/utils';
 
+import { sendVerificationEmail } from '../../../helpers/email.js';
 import { asyncWrapper } from '../../../utils/asyncWrapper.js';
 
 import type { GetEmailByExpiredToken } from '@nangohq/types';
