+++ packages/server/lib/controllers/v1/account/postForgotPassword.ts
@@ -1,9 +1,10 @@
 import jwt from 'jsonwebtoken';
 import * as z from 'zod';
 
-import { sendResetPasswordEmail, userService } from '@nangohq/shared';
+import { userService } from '@nangohq/shared';
 import { requireEmptyQuery, zodErrorToHTTP } from '@nangohq/utils';
 
+import { sendResetPasswordEmail } from '../../../helpers/email.js';
 import { asyncWrapper } from '../../../utils/asyncWrapper.js';
 import { resetPasswordSecret } from '../../../utils/utils.js';
