+++ packages/server/lib/controllers/v1/account/resendVerificationEmailByEmail.ts
@@ -1,8 +1,9 @@
 import * as z from 'zod';
 
-import { sendVerificationEmail, userService } from '@nangohq/shared';
+import { userService } from '@nangohq/shared';
 import { requireEmptyQuery, zodErrorToHTTP } from '@nangohq/utils';
 
+import { sendVerificationEmail } from '../../../helpers/email.js';
 import { asyncWrapper } from '../../../utils/asyncWrapper.js';
 
 import type { ResendVerificationEmailByEmail } from '@nangohq/types';
