+++ packages/server/package.json
@@ -45,6 +45,7 @@
         "exponential-backoff": "3.1.1",
         "express": "5.1.0",
         "express-session": "1.18.2",
+        "he": "^1.2.0",
         "helmet": "7.1.0",
         "jsonwebtoken": "9.0.2",
         "lodash-es": "4.17.21",
