+++ packages/shared/package.json
@@ -20,9 +20,7 @@
     "dependencies": {
         "@aws-sdk/client-s3": "3.796.0",
         "@hapi/boom": "10.0.1",
-        "@nangohq/account-usage": "file:../account-usage",
         "@nangohq/database": "file:../database",
-        "@nangohq/email": "file:../email",
         "@nangohq/kvstore": "file:../kvstore",
         "@nangohq/nango-yaml": "file:../nango-yaml",
         "@nangohq/providers": "file:../providers",
