+++ packages/shared/tsconfig.json
@@ -6,9 +6,6 @@
     },
     "references": [
         {
-            "path": "../account-usage"
-        },
-        {
             "path": "../database"
         },
         {
@@ -34,9 +31,6 @@
         },
         {
             "path": "../providers"
-        },
-        {
-            "path": "../email"
         }
     ],
     "include": ["lib/**/*", "../utils/lib/vitest.d.ts"]
